styleupapp/requirements.txt
dash==0.36.0
dash-core-components==0.43.0
dash-html-components==0.13.5
dash-renderer==0.17.0
dash-table==3.7.0
dask==1.1.4
Flask==1.0.1
Flask-Caching==1.4.0
Flask-Compress==1.4.0
future==0.17.1
html5lib==1.0.1
numpy==1.17.4
numpydoc==0.8.0
openpyxl==2.6.1
packaging==19.0
pandas==0.24.2
pandas-bokeh==0.2
pickleshare==0.7.5
Pillow==6.0.0
scikit-image==0.14.2
scikit-learn==0.21.3
scipy==1.2.1
matplotlib==3.0.2
gunicorn
opencv-python==4.1.1.26
